debut couleur

+++ headers/struct.h
@@ -6,7 +6,7 @@
 /*   By: astachni <astachni@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 17:15:02 by astachni          #+#    #+#             */
-/*   Updated: 2023/10/05 18:24:09 by astachni         ###   ########.fr       */
+/*   Updated: 2023/10/09 18:12:34 by astachni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_fpoint
 {
 	float	x;
 	float	y;
-	float	distance;
+	double	distance;
 	float	height;
 }	t_fpoint;
